Extract oneshot unit test from holdtap
